# api.yml
# This is the OpenAPI specification for your service. It is used to generate the client code.
# To generate the server code, use the Gradle task `openApiGenerate`. The 
# generated code will be placed in `build/generate-resources/main`.
#
# We will evaluate you based on how well you design your API.
# 1. How well it follows REST principles.
# 2. How easy it is to understand and use.
#
# References
# 1. https://swagger.io/specification/
openapi: "3.0.0"
info:
  version: 1.0.0
  title: User Service
  license:
    name: MIT
servers:
  - url: http://localhost
paths:
  /users/registration:
    post:
      operationId: userRegistration
      requestBody:
        description: Registration payload
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegistrationRequest'
      responses:
        '200':
          description: Successfully registration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegistrationResponse'
        '400':
          description: Failed registration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegistrationResponse'
  /users/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: The user identifier
        schema:
          type: integer
          format: int64
    get:
      operationId: getUser
      responses:
        '200':
          description: The user being returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetUserResponse'
        '404':
          description: User not exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetUserResponse'
    patch:
      operationId: updateUser
      requestBody:
        description: Update user payload
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: The user being updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateUserResponse'
        '404':
          description: User not exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateUserResponse'
        '409':
          description: Phone number already in use
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateUserResponse'
components:
  schemas:
    RegistrationRequest:
      type: object
      required: [ fullName, countryCode, phoneNumber, password ]
      properties:
        fullName:
          type: string
          description: Must be at minimum 3 characters and maximum 60 characters
          x-oapi-codegen-extra-tags:
            validate: required,min=3,max=60
        countryCode:
          type: string
          description: Phone country code
          example: '+62'
          x-oapi-codegen-extra-tags:
            validate: required,min=3,max=7
        phoneNumber:
          type: string
          description: Must be at minimum 10 characters and maximum 13 characters
          x-oapi-codegen-extra-tags:
            validate: required,numeric,min=10,max=13
        password:
          type: string
          description: Must be minimum 6 characters and maximum 64 characters, containing at least 1 capital characters AND 1 number AND 1 special (non alphanumeric) characters
          x-oapi-codegen-extra-tags:
            validate: required,min=6,max=64
#            validate: required,min=6,max=64,regexp=^(?=.*?[A-Z])(?=.*?[0-9])(?=.*?[#?!@$%^&*-])
    RegistrationResponse:
      type: object
      required: [code, data, errors]
      properties:
        code:
          type: string
        data:
          type: object
          properties:
            id:
              type: integer
              format: int64
        errors:
          type: array
    GetUserResponse:
      type: object
      required: [code, data, errors]
      properties:
        code:
          type: string
        data:
          type: object
          properties:
            id:
              type: integer
              format: int64
            fullName:
              type: string
            phoneNumber:
              type: string
        errors:
          type: array
    UpdateUserRequest:
      type: object
      required: [ fullName, countryCode, phoneNumber ]
      properties:
        fullName:
          type: string
          description: Must be at minimum 3 characters and maximum 60 characters
          x-oapi-codegen-extra-tags:
            validate: omitempty,min=3,max=60
        countryCode:
          type: string
          description: Phone country code
          example: '+62'
          x-oapi-codegen-extra-tags:
            validate: omitempty,min=3,max=7
        phoneNumber:
          type: string
          description: Must be at minimum 10 characters and maximum 13 characters
          x-oapi-codegen-extra-tags:
            validate: omitempty,required_with=CountryCode,numeric,min=10,max=13
    UpdateUserResponse:
      type: object
      required: [code, data, errors]
      properties:
        code:
          type: string
        data:
          type: object
          properties:
            id:
              type: integer
              format: int64
            fullName:
              type: string
            phoneNumber:
              type: string
        errors:
          type: array